optimize installation
