+++ src/install.rs
@@ -1,125 +1,18 @@
+use crate::manage_utils;
 use crate::args::Args;
 use log::{error, info, warn};
 use std::fs::File;
-use std::io::{Read, Write};
+use std::io::Write;
 use std::process::{exit, Command};
-use std::{env, fs, io};
-
-#[derive(Debug)]
-pub enum PID1 {
-    Systemd,
-    OpenRC,
-}
-pub fn check_pid1() -> PID1 {
-    let pid1_path = "/proc/1/comm";
-    let mut pid1_file = match File::open(pid1_path) {
-        Ok(file) => file,
-        Err(e) => {
-            error!("无法获取守护进程信息: {}", e);
-            exit(1);
-        }
-    };
-    let mut string_of_pid1 = String::new();
-    match pid1_file.read_to_string(&mut string_of_pid1) {
-        Ok(_) => {}
-        Err(e) => {
-            error!("无法读取 /proc/1/comm: {}", e);
-            exit(1);
-        }
-    };
-
-    let pid1 = if string_of_pid1 == "systemd\n" {
-        PID1::Systemd
-    } else if string_of_pid1 == "openrc\n" || string_of_pid1 == "init\n" {
-        PID1::OpenRC
-    } else {
-        error!("无法识别守护进程, 退出！");
-        exit(1);
-    };
-
-    println!("检测到守护进程: {:?}", pid1);
-
-    pid1
-}
+use std::{fs, io};
 
 pub fn install_to_systemd(args: Args) {
-    // 检查操作系统是否为 Linux
-    if env::consts::OS != "linux" {
-        error!("Install 功能仅适用于 Linux 系统");
-        exit(1);
-    }
 
-    if env::var("USER") == Ok("root".to_string()) {
-        info!("正在使用 root 用户");
-    } else {
-        error!("非 root 用户, 请使用 root 用户运行 Install 功能");
-        exit(1);
-    }
+    manage_utils::check_installed("/etc/systemd/system/akile_monitor_client.service");
+    manage_utils::copy_binary();
 
-    // 检查是否已存在相同名称的服务文件
-    match fs::metadata("/etc/systemd/system/akile_monitor_client.service") {
-        Ok(_) => {
-            error!("已存在相同名称的服务文件, 请先使用 `--uninstall` 参数卸载后再安装");
-            exit(1);
-        }
-        Err(_) => {}
-    }
-
-    // 复制可执行文件到 /usr/bin
-    match env::current_exe() {
-        Ok(path_to_bin) => {
-            if path_to_bin.to_str().unwrap() == "/usr/bin/ak_monitor_client_rs" {
-                info!("无需复制可执行文件");
-            } else {
-                match fs::copy(path_to_bin, "/usr/bin/ak_monitor_client_rs") {
-                    Ok(_) => {
-                        info!("成功将可执行文件复制到 /usr/bin/ak_monitor_client_rs");
-                    }
-                    Err(e) => {
-                        error!(
-                            "无法将可执行文件复制到 /usr/bin/ak_monitor_client_rs: {}",
-                            e
-                        );
-                        exit(1);
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            error!("无法获取可执行文件路径: {}", e);
-            exit(1);
-        }
-    }
-
-    let template = r#"[Unit]
-Description=Akile Monitor Client Service
-After=network.target
-
-[Install]
-WantedBy=multi-user.target
-
-[Service]
-Type=simple
-ExecStart=COMMAND
-Restart=always
-"#;
-    let debug_ = if args.debug { "--debug" } else { "" };
-
-    let tls = if args.tls { "--tls true" } else { "--tls false" };
-
-    let command = format!(
-        "/usr/bin/ak_monitor_client_rs {} {} -n \"{}\" -s \"{}\" -a \"{}\" -i {} -f {} --monitor-path \"{}\"",
-        debug_,
-        tls,
-        args.name,
-        args.server,
-        args.auth_secret,
-        args.interval,
-        args.fake_times,
-        args.monitor_path
-    );
-
-    let service_string = template.replace("COMMAND", &command);
+    let command_args = manage_utils::generate_client_args(args);
+    let service_string = SERVICE_TEMPLATE_SYSTEMD.replace("COMMAND_ARGS", &command_args);
     info!("最终服务文件: ");
     println!("{}", service_string);
 
@@ -218,13 +111,6 @@ Restart=always
 pub fn uninstall_from_systemd() {
     info!("开始卸载 Akile Monitor Client Service");
 
-    if env::var("USER") == Ok("root".to_string()) {
-        info!("正在使用 root 用户");
-    } else {
-        error!("非 root 用户, 请使用 root 用户运行 Uninstall 功能");
-        exit(1);
-    }
-
     match Command::new("systemctl")
         .arg("stop")
         .arg("akile_monitor_client.service")
@@ -284,113 +170,17 @@ pub fn uninstall_from_systemd() {
             warn!("无法运行 systemctl daemon-reload: {}", e);
         }
     }
-    match fs::remove_file("/usr/bin/ak_monitor_client_rs") {
-        Ok(_) => {
-            info!("成功删除 /usr/bin/ak_monitor_client_rs");
-        }
-        Err(e) => {
-            warn!("无法删除 /usr/bin/ak_monitor_client_rs: {}", e);
-        }
-    }
+    manage_utils::delete_binary();
     info!("成功卸载 Akile Monitor Client Service");
     exit(1);
 }
 
 pub fn install_to_openrc(args: Args) {
-    // 检查操作系统是否为 Linux
-    if env::consts::OS != "linux" {
-        error!("Install 功能仅适用于 Linux 系统");
-        exit(1);
-    }
+    manage_utils::check_installed("/etc/init.d/akile_monitor_client_rs");
+    manage_utils::copy_binary();
 
-    if env::var("USER") == Ok("root".to_string()) {
-        info!("正在使用 root 用户");
-    } else {
-        error!("非 root 用户, 请使用 root 用户运行 Install 功能");
-        exit(1);
-    }
-
-    // 检查是否已存在相同名称的服务文件
-    match fs::metadata("/etc/init.d/akile_monitor_client_rs") {
-        Ok(_) => {
-            error!("已存在相同名称的服务文件, 请先使用 `--uninstall` 参数卸载后再安装");
-            exit(1);
-        }
-        Err(_) => {}
-    }
-
-    // 复制可执行文件到 /usr/bin
-    match env::current_exe() {
-        Ok(path_to_bin) => {
-            if path_to_bin.to_str().unwrap() == "/usr/bin/ak_monitor_client_rs" {
-                info!("无需复制可执行文件");
-            } else {
-                match fs::copy(path_to_bin, "/usr/bin/ak_monitor_client_rs") {
-                    Ok(_) => {
-                        info!("成功将可执行文件复制到 /usr/bin/ak_monitor_client_rs");
-                    }
-                    Err(e) => {
-                        error!(
-                            "无法将可执行文件复制到 /usr/bin/ak_monitor_client_rs: {}",
-                            e
-                        );
-                        exit(1);
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            error!("无法获取可执行文件路径: {}", e);
-            exit(1);
-        }
-    }
-
-    let template = r#"#!/sbin/openrc-run
-
-command=/usr/bin/ak_monitor_client_rs
-command_args="COMMAND_ARGS_HERE"
-pidfile=/run/akile_monitor_client_rs.pid
-description="Akile Monitor Slave Client (Rust) Service"
-
-depend() {
-    need net
-    after firewall
-}
-
-start() {
-    ebegin "Starting Akile Monitor Slave Client (Rust) Service"
-    start-stop-daemon --start --background --make-pidfile --pidfile $pidfile --exec $command -- $command_args
-    eend $?
-}
-
-stop() {
-    ebegin "Stopping Akile Monitor Slave Client (Rust) Service"
-    start-stop-daemon --stop --pidfile $pidfile
-    eend $?
-}
-
-restart() {
-    stop
-    start
-}
-"#;
-    let debug_ = if args.debug { "--debug" } else { "" };
-
-    let tls = if args.tls { "--tls true" } else { "--tls false" };
-
-    let command = format!(
-        "{} {} -n \"{}\" -s \"{}\" -a \"{}\" -i {} -f {} --monitor-path \"{}\"",
-        debug_,
-        tls,
-        args.name,
-        args.server,
-        args.auth_secret,
-        args.interval,
-        args.fake_times,
-        args.monitor_path
-    );
-
-    let service_string = template.replace("COMMAND_ARGS_HERE", &command);
+    let command_args = manage_utils::generate_client_args(args);
+    let service_string = SERVICE_TEMPLATE_OPENRC.replace("COMMAND_ARGS", &command_args);
     info!("最终服务文件: ");
     println!("{}", service_string);
 
@@ -428,6 +218,25 @@ restart() {
         Err(e) => {
             warn!(
                 "无法设置 /etc/init.d/akile_monitor_client_rs 为可执行文件: {}",
+                e
+            );
+        }
+    }
+
+    match Command::new("rc-update")
+        .arg("-u")
+        .output()
+    {
+        Ok(tmp) => {
+            if tmp.status.success() {
+                info!("成功重载 OpenRC 服务");
+            } else {
+                warn!("无法重载 OpenRC 服务")
+            }
+        }
+        Err(e) => {
+            warn!(
+                "无法重载 OpenRC 服务: {}",
                 e
             );
         }
@@ -492,13 +301,6 @@ restart() {
 pub fn uninstall_from_openrc() {
     info!("开始卸载 Akile Monitor Client Service");
 
-    if env::var("USER") == Ok("root".to_string()) {
-        info!("正在使用 root 用户");
-    } else {
-        error!("非 root 用户, 请使用 root 用户运行 Uninstall 功能");
-        exit(1);
-    }
-
     match Command::new("rc-service")
         .arg("akile_monitor_client_rs")
         .arg("stop")
@@ -545,14 +347,51 @@ pub fn uninstall_from_openrc() {
         }
     }
 
-    match fs::remove_file("/usr/bin/ak_monitor_client_rs") {
-        Ok(_) => {
-            info!("成功删除 /usr/bin/ak_monitor_client_rs");
-        }
-        Err(e) => {
-            warn!("无法删除 /usr/bin/ak_monitor_client_rs: {}", e);
-        }
-    }
+    manage_utils::delete_binary();
+
     info!("成功卸载 Akile Monitor Client Service");
     exit(1);
 }
+
+static SERVICE_TEMPLATE_OPENRC: &str = r#"#!/sbin/openrc-run
+
+command=/usr/bin/ak_monitor_client_rs
+command_args="COMMAND_ARGS"
+pidfile=/run/akile_monitor_client_rs.pid
+description="Akile Monitor Client Service"
+
+depend() {
+    need net
+    after firewall
+}
+
+start() {
+    ebegin "Starting Akile Monitor Client Service"
+    start-stop-daemon --start --background --make-pidfile --pidfile $pidfile --exec $command -- $command_args
+    eend $?
+}
+
+stop() {
+    ebegin "Stopping Akile Monitor Client Service"
+    start-stop-daemon --stop --pidfile $pidfile
+    eend $?
+}
+
+restart() {
+    stop
+    start
+}
+"#;
+
+const SERVICE_TEMPLATE_SYSTEMD: &str = r#"[Unit]
+Description=Akile Monitor Client Service
+After=network.target
+
+[Install]
+WantedBy=multi-user.target
+
+[Service]
+Type=simple
+ExecStart=/usr/bin/ak_monitor_client_rs COMMAND_ARGS"
+Restart=always
+"#;
